refactor: don't "sanitize" backslashes for display
I added this in 7f5457b058488 ("fix: don't let control characters mess
up the display", 2022-08-31) but it turns out it is a bit overzealous.

We don't want to render `\r` as `^M`, sure, but we are about to build
the Lua version of `:CommandTSearch`, and we want to render stuff like
`\vfoo` as `\vfoo` and not as `\\vfoo`.

We can remove the backslash escaping. This is a display-only thing, so
it won't affect how selections are actually opened.