feat: port `:CommandTJump` to Lua
So, the Ruby version parsed the output of `:jumps` and I thought I'd be
clever here by doing it with `getjumplist()` instead. I am _kind of_
convinced that it works acceptably.

How stuff works: depending on your `'shada'` setting, your jumplist may
be persisted and restored across Neovim restarts. So, if you land in
Neovim and run `:ls` you'll see a single unnamed buffer. If you run
`:ls!` to see unlisted buffers too, you'll still only see that one
buffer.

But if you run `:jumps`, you'll see something curious. Not only will you
see the places you jumped to in previous sessions, but you'll see that
Neovim actually created unlisted buffers for those files, visible with
`:ls!`. Personally, I find this mildly surprising.

In any case, the way this interacts with the code in this commit is
interesting. Instead of calling `:jumps`, which has the above behavior,
we call `getjumplist()`. It turns out that this has the same behavior,
and it's useful to us. We do this, more or less:

1. Loop over all tabpages.
2. Loop over all windows inside each tabpage.
3. Loop over all jumplist entries inside each window.
4. If there is no `filename`, look one up from the `bufnr`.
5. Make sure the file really exists on disk.

In my experience, _no_ entries have `filename` values, and the list only
contains jump list entries for loaded buffers. But, as noted above, and
shown by `:ls!`, even those buffers from prior sessions end up being
loaded, so this actually works.

One final thing to note: is it worth doing the "loop over tabpages, loop
over windows" thing? It may be overkill, to be honest, because in
practice, when you create a new window, the jumplist ends up getting
copied, so, pretty much every jumplist ends up dragging with it the
history of everything that came before it. Well, not everything, but 100
entries' worth...

From `:help jumplist`:

> There is a separate jump list for each window. The maximum number of
> entries is fixed at 100.
>
> ...
>
> When you split a window, the jumplist will be copied to the new
> window.

Also relevant, `:help 'jumpoptions':

                                                    *'jumpoptions'* *'jop'*
    'jumpoptions' 'jop'     string  (default "clean")
                            global
            List of words that change the behavior of the |jumplist|.
              stack         Make the jumplist behave like the tagstack.
                            Relative location of entries in the jumplist is
                            preserved at the cost of discarding subsequent entries
                            when navigating backwards in the jumplist and then
                            jumping to a location.  |jumplist-stack|

              view          When moving through the jumplist, |changelist|,
                            |alternate-file| or using |mark-motions| try to
                            restore the |mark-view| in which the action occurred.

              clean         Remove unloaded buffers from the jumplist.
                            EXPERIMENTAL: this flag may change in the future.

"clean" is the default, but it may be that keeping it empty is actually
more useful (because more things will remain in the list). On the
flipside, the bigger the list, the greater the cost of the
in-the-background loading into unlisted buffers...