feat: port `:CommandTTag` to Lua
Note that this has some limitations, just like the old version had. The
main problem is that you can't tell `:tag` _which_ tag to jump to, even
when displaying multiple filenames.

I did consider trying to hack around this by exploiting
`|tag-priority|`:

                                                            *tag-priority*
    When there are multiple matches for a tag, this priority is used:
    1. "FSC"  A full matching static tag for the current file.
    2. "F C"  A full matching global tag for the current file.
    3. "F  "  A full matching global tag for another file.
    4. "FS "  A full matching static tag for another file.
    5. " SC"  An ignore-case matching static tag for the current file.
    6. "  C"  An ignore-case matching global tag for the current file.
    7. "   "  An ignore-case matching global tag for another file.
    8. " S "  An ignore-case matching static tag for another file.

ie. we could switch to the file in question, _then_ run `:tag`. But that
won't always work; in the case where there is only one matching
tag in the file, it will go to the right place, but if there are
multiple matching tags in a single file, it will jump to the first one,
which may not be the one the user wanted. The user can still get to the
one they want by jumping with `:tn`, but it's not an ideal user
experience.

A more robust solution will require us to thread through the `cmd` info
in the `taglist()` result and use that to jump to the right place within
the file. Here's a two sample entries, for example, for Command-T itself
(the tags file in question generated with `ctags -R`), for jumping to
the `scanner_free()` definition:

  {
    'cmd': '/^void scanner_free(scanner_t *scanner) {$/',
    'static': 0,
    'name': 'scanner_free',
    'kind': 'f',
    'filename': 'lua/wincent/commandt/lib/scanner.c'
  }, {
    'cmd': '/^#define scanner_free /',
    'static': 0,
    'name': 'scanner_free',
    'kind': 'd',
    'filename': 'lua/wincent/commandt/lib/scanner.h'
  }

Those correspond to these entries in the `tags` file:

    scanner_free	lua/wincent/commandt/lib/scanner.c	/^void scanner_free(scanner_t *scanner) {$/;"	f
    scanner_free	lua/wincent/commandt/lib/scanner.h	/^#define scanner_free /;"	d

These are in two separate files, so we caen reliably jump to them. But
if they were in the same file, we could not. What we'd need to do is one
of the following:

- Allow the list scanner (and probably only the list scanner, because
  it would be too complicated with the command scanner etc) to take not
  just a list of strings, but a list of tables. The table would contain
  what is shown to the user (ie. the haystack they search through), but
  it would also contain the extra metadata that we'll need in order to
  open the exact result that they selected. This table would need to be
  threaded though an unpleasant number of places, so that by the time we
  get to the `open()` callback, we have all we need to reconstruct the
  user's intent.
- Do the same without all the threading. Assume that candidate strings
  are unique, use them as keys into a `context` data structure that we
  pass back into the `open()` call back, and look up the `cmd` from the
  context.

As that may be a bit involved, I am going to do it in a separate commit.