refactor: rename ex-command `kind` to `ex_command`
Continues work started in parent commit, to reduce the number of places
we use the word `kind` to refer to different things.

Now the main use is in the options spec.

There is also a use in the tests, where a context can be of `kind`
`'context'`, `'test'`, or `'skipped'`, but that is separate enough that
I am not worried about people confusing it too much, so I will leave it
as-is rather than renaming it to something like `context_kind` or
`context_type` (although I may change my mind if I think of a really
great name for it).