perf: limit based on available height as opposed to configured height
If the user passes in a height of 10,000, we'll never be able to render
that many results, so telling the matcher to gather that many is a waste
of resources.

I don't want to add a reference to `vim.o.lines` in the `lib` layer, so
instead pass it in via a new `context` parameter. This leads to
duplication because there are 4 call sites (not counting tests), but I
prefer the duplication over letting a `vim` reference leak into the
`lib` layer. I could have extracted this into an auxiliary wrapper
function that is called from the four places, but that doesn't seem
worth it; I'd have to come up with a name, and naming is hard, and I
don't think I'm very likely to add more call sites in the future.

The other thing I'm fixing here is the `merge` call which takes a
well-formed `options` object and uses it to overwrite a set of default
values (fine) plus one value (`limit`) which is not part of the options
schema (not fine). It then takes the merged object, which has a lot of
keys, and plucks out a subset that actually get forwarded.

So, instead of that, add a `fetch()` function that we use to extract
just the values we want. There is no `merge()` call and no intermediate
object with keys/values we don't care about.

Remaining smell is that we start with `any`-typed `ignore_case` and
`ignore_spaces` variables (in practice, booleans or functions), and we
end up overwriting them via two `if` statements, changing the type in
that case to boolean. Outside they `if`s, though, they remain `any`. So
that's all a bit meh.