refactor: remove unused "recurse" setting
We no longer expose a proper setting here. It is always `true` in
practice, except in the benchmarks where we offered an override via the
environment.

If there is ever a call for this in the future, we can resurrect it, but
for now, leaving it out presents us with an opportunity to simplify.
It may even be a tiny bit faster (1.3% better CPU time, and 2.4%
better wall time), with reasonable confidence, due to saving us some
conditional checks. I'm a bit skeptical that it really makes that much
of a difference, but the numbers are what they are:

    Summary of cpu time and (wall time):

                        best    avg      sd     +/-      p     (best)    (avg)      (sd)     +/-     p
         pathological 0.19878 0.20760 0.06574 [-8.0%]  0.005 (0.19877) (0.20760) (0.06573) [-8.0%] 0.005
            command-t 0.17290 0.17914 0.05285 [-0.6%]        (0.17289) (0.17914) (0.05286) [-0.6%]
    chromium (subset) 1.34179 1.35355 0.02208 [-1.0%] 0.0005 (0.28701) (0.29202) (0.01367) [-0.8%] 0.025
     chromium (whole) 1.14923 1.15236 0.00913 [-1.2%] 0.0005 (0.12593) (0.12894) (0.00875) [-1.8%] 0.005
           big (400k) 1.72400 1.73176 0.01878 [-1.0%] 0.0005 (0.18604) (0.18899) (0.00862) [-0.8%] 0.025
                total 4.60219 4.62441 0.10340 [-1.3%] 0.0005 (0.97700) (0.99670) (0.10262) [-2.4%] 0.005