perf: use voodoo coding to eek out minor performance increase
Looking at 6% improvement in CPU time and 1.5% improvement in wall time:

    Summary of cpu time and (wall time):

                        best    avg      sd     +/-      p     (best)    (avg)      (sd)     +/-      p
         pathological 0.21343 0.22251 0.06979 [+6.6%] 0.0005 (0.21343) (0.22251) (0.06977) [+6.5%] 0.0005
            command-t 0.16698 0.17370 0.06094 [-2.9%] 0.0005 (0.16698) (0.17375) (0.06171) [-2.9%] 0.0005
    chromium (subset) 1.28794 1.29834 0.03082 [-5.0%] 0.0005 (0.28016) (0.28721) (0.01598) [-1.6%]  0.005
     chromium (whole) 1.06593 1.06970 0.01175 [-7.8%] 0.0005 (0.11729) (0.12135) (0.01042) [-6.2%] 0.0005
           big (400k) 1.60554 1.60981 0.01459 [-7.6%] 0.0005 (0.17389) (0.17605) (0.00712) [-7.1%] 0.0005
                total 4.34685 4.37406 0.13084 [-6.0%] 0.0005 (0.95780) (0.98088) (0.11848) [-1.5%] 0.0005

Why does this work? TBH, it's hard to be 100% certain without having
instruction-level profiling, and I don't have that. At first glance, it
looks like we're delaying the check for memoized result and potentially
doing _more_ work than before. But if the benchmark is to be believed,
this might actually be for the best. Perhaps, the operations involved in
looking up the memoization structure are more expensive than you might
expect. In code like this:

    memoized = &m->memo[j * m->needle_length + i];
    if (*memoized != UNSET_SCORE) {
        return *memoized > seen_score ? *memoized : seen_score;
    }

we are getting the address (`&`) of a float in an array, having
calculated the index of that float by following a struct pointer (`->`)
and doing a multiplication (`*`) and an addition (`+`); we're also doing a
comparison (`!=`) and loading the actual value from the address (`*`) at
least once (in the `if` condition) and possibly more than once (in the
ternary).

In contrast, in the new code, we're only doing all of that in the
relatively rare scenarios that:

- The current haystack character is a dot at the start of the haystack
  or right after a slash (ie. the haystack represents a dot-file or a
  file with a dot-directory ancestor), and we're not searching for a
  dot, and the user doesn't have the `always_show_dot_files` setting
  turned on.
- The current haystack character matches the current needle character
  (most of the time, it won't).

This analysis is born out by the fact that the "pathological" workload
did get slower in the benchmarks above; this is because in that
workload, the second "rare" scenario isn't rare at all. In that case,
instead of doing the expensive memoization look-up all the time, but
bailing immediately as a result, we're now doing some additional work,
then doing the expensive memoization look-up, then bailing. But given
that most workloads get faster, this is overall a win.

Now, I don't _think_ this behavior changes the results of the scoring
algorithm. As you can see, the motivation for all this was to print the
memoization struct so I could add it to a comment in the tests. The
tests still pass, and the printed struct has the same contents before
and after the memoization changes.