perf: avoid some pointer indirection
Apparently, worth 4.5% of CPU time and 2.8% of wall time:

    Summary of cpu time and (wall time):

                        best    avg      sd     +/-      p     (best)    (avg)      (sd)     +/-      p
         pathological 0.20729 0.21859 0.07567 [-1.9%]  0.005 (0.20737) (0.21863) (0.07571) [-1.9%]  0.005
            command-t 0.15597 0.16216 0.05334 [-2.0%] 0.0005 (0.15600) (0.16218) (0.05335) [-2.0%] 0.0005
    chromium (subset) 1.21716 1.22648 0.02489 [-3.5%] 0.0005 (0.26888) (0.27404) (0.01121) [-2.3%] 0.0005
     chromium (whole) 0.96999 0.97436 0.01232 [-5.3%] 0.0005 (0.10804) (0.11196) (0.01187) [-3.9%] 0.0005
           big (400k) 1.46021 1.46560 0.02028 [-5.4%] 0.0005 (0.15898) (0.16143) (0.00690) [-5.2%] 0.0005
                total 4.01989 4.04719 0.11325 [-4.5%] 0.0005 (0.90571) (0.92824) (0.11004) [-2.8%] 0.0005

What are the cumulative effects of some of recent optimizations? Let's
compare with b0bf414eb0b348842 ("perf: make benchmarks more realistic",
2022-07-12). At a glance, that looks to be the farthest back we can go
without the benchmarks themselves having substantive changes. The format
then was different, so I went back and checked out the commit, rebuilt,
and here is the last run's numbers:

    Final----------------      total         wall
    pathological            1.075292   (1.075315)
    command-t               0.620796   (0.620811)
    chromium (subset)       4.072396   (0.640318)
    chromium (whole)        5.114491   (0.467661)
    big (400k)              7.629578   (0.671459)

It doesn't tally the numbers so I'll to that here, so we can compare:

    total                   18.512653  (3.475564)

So, unless something escapes me, it looks like we've gotten about 4.6x
faster for CPU time, and about 3.8x faster for wall time in the last
three years.

Fun fact: the numbers reported in the commit message for
b0bf414eb0b348842 also tell a story of how much faster my machine has
gotten in that time; here they are, for comparison:

    Final----------------      total         wall
    pathological            1.170620   (1.170639)
    command-t               0.759809   (0.759821)
    chromium (subset)       4.977437   (0.950148)
    chromium (whole)        7.021354   (1.011983)
    big (400k)              10.533377  (1.455907)

Again, the tally:

    total                   24.462597  (5.348498)

So, not as impressive gains, but gains nonetheless: a 25% improvement
in CPU time, and a 35% improvement in wall time.